Filters out route tables that have None for id
I'm not really sure how this happens. I changed the main route table of
my VPC and added a new one and now there's an additional route table
that appears in boto that doesn't seem to exist. It is likely the bug is
in boto, so this is more of a patch.