Removes IgnoreDefaultRoute from chrome ONC
We had discussed this on 10 September 2018 and a patch was put through,
and I confirmed it had fixed our issues after it went live, but this
week it occurred to me I had tested my own code and not the patch you'd
put through. As such, full tunnel connections still weren't working
correctly. We realized this after upgrading to the latest Pritunl. This
patch is what is on our Pritunl web host now.

On September 13th, you said:
> I don't think it will work but there isn't a significant use case for
> routing all traffic on Chrome OS so it isn't something that I plan on
> fixing.

You were ultimately correct, since the patch didn't follow the
instructions Google had given us:

> On client side you need to remove
> "IgnoreDefaultRoute": true,
> Line from ONC file.

Instead of removing the line, the previous patch changed `true` to
`false`. This patch removes the line altogether, which I can confirm
fixes the problem.