Fixed invalid ChromeOS profile generation
When trying to import an .onc (ChromeOS profile) file, the profile
imports successfully but ChromeOS complains that one of the user
certificates is invalid. Upon closer investigation, it appears that the
base64 representation was substituted incorrectly into the template,
presumably missed out when Pritunl introduced Python 3 compatibility.

This commit runs .decode() on the return value of b64encode so as to
ensure a string gets substituted instead of bytes.